\name{gen.net}
\alias{gen.net}
\title{Generate a Laplace Matrix}
\description{
   Generate a Laplace Matrix.
}
\usage{
gen.net(p, nc = 10, pc = 0.4, pcc = 0.13, psn = 0.04, normalize = FALSE, plot.graph = FALSE)
}
\arguments{
  \item{p}{ }
  \item{nc}{ }
  \item{pc}{ }
  \item{pcc}{ }
  \item{psn}{ }
  \item{normalize}{ }
  \item{plot.graph}{ }
}
\value{
 \item{L}{ }
 \item{A}{ }
}
\references{
  Wen, C., Zhang, A., Quan, S. and Wang, X. (2017) \emph{BeSS: an R package for best subset selection in linear, logistic and CoxPH models}. \url{https://arxiv.org/abs/1709.06254}.\cr
  Wen, C., Zhu, J. and Wang, X. (2018) \emph{Best Subset Selection in Reduced Rank Regression}, Technical Report.
}

\author{
  Canhong Wen, Shijie Quan and Xueqin Wang\cr
  Maintainer: Canhong Wen \email{wench@ustc.edu.cn}
}

